backend/requirements.txt
Django
djangorestframework
django-cors-headers
gunicorn
pyfiglet
python-dotenv
whitenoise
requests